# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PartyNest API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Submit a quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Quote request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"

    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List quote requests (for business)
      parameters:
        - name: businessId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get user's wishlist
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"

    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add a business to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWishlistRequest"
      responses:
        "201":
          description: Added to wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{businessId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove a business from wishlist
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed from wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Business:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        locationType:
          type: string
          enum: [indoor, outdoor, both]
        minAge:
          type: integer
        maxAge:
          type: integer
        capacity:
          type: integer
        basePrice:
          type: number
        photos:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        rules:
          type: string
        phone:
          type: string
        website:
          type: string
        visibilityPoints:
          type: integer
        views:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, ownerId, name, city, address, locationType, capacity, basePrice, createdAt]
...
    Quote:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        eventDate:
          type: string
          format: date
        guestCount:
          type: integer
        childrenAge:
          type: string
        message:
          type: string
        services:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, responded, closed]
        business:
          $ref: "#/components/schemas/Business"
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, name, email, guestCount, status, createdAt]
    CreateQuoteRequest:
      type: object
      properties:
        businessId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        eventDate:
          type: string
          format: date
        guestCount:
          type: integer
        childrenAge:
          type: string
        message:
          type: string
        services:
          type: array
          items:
            type: string
      required: [businessId, name, email, guestCount]
...
    WishlistItem:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessId:
          type: string
        business:
          $ref: "#/components/schemas/Business"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, businessId, createdAt]
    AddWishlistRequest:
      type: object
      properties:
        businessId:
          type: string
      required: [businessId]
...